python_scraper/requirements.txt
playwright==1.52.0
httpx==0.28.1
python-dotenv==1.1.0
beautifulsoup4==4.13.4
tqdm==4.67.1
